Removed copy construction and assignment for nodes, added Image class.
The testpdf example was updated to include an image.
No unit tests for the Image class have been written yet.

+++ NodeCore/graphics/Utils.h
@@ -26,12 +26,12 @@ namespace NodeCore {
 
 inline float radians(float degrees)
 {
-    return degrees * M_PI / 180;
+    return (float) (degrees * M_PI / 180);
 }
 
 inline float degrees(float radians)
 {
-    return radians * 180 / M_PI;
+    return (float) (radians * 180 / M_PI);
 }
 
 }
